changed return value of mpfr_cmp2
[From SVN r1047 (trunk)]

+++ mpfr-impl.h
@@ -88,7 +88,7 @@ int mpfr_can_round_raw _PROTO ((mp_limb_t *, mp_prec_t, int, mp_prec_t,
 				mp_rnd_t, mp_rnd_t, mp_prec_t));
 double mpfr_get_d2 _PROTO ((mpfr_srcptr, long)); 
 mp_size_t mpn_sqrtrem_new _PROTO ((mp_limb_t *, mp_limb_t *, mp_limb_t *, mp_size_t));
-int mpfr_cmp2 _PROTO ((mpfr_srcptr, mpfr_srcptr));
+unsigned long mpfr_cmp2 _PROTO ((mpfr_srcptr, mpfr_srcptr));
 long _mpfr_ceil_log2 _PROTO ((double));
 long _mpfr_floor_log2 _PROTO ((double));
 double _mpfr_ceil_exp2 _PROTO ((double));
